2.82v

+++ lhc_web/lib/models/lhchat/erlhcoreclassmodelchatonlineuser.php
@@ -542,37 +542,49 @@ class erLhcoreClassModelChatOnlineUser
         }
     }
 
-    public static function cleanupOnlineUsers()
+    public static function cleanupOnlineUsers($params = array())
     {
-        $db = ezcDbInstance::get();
+        $cleanupCronjob = erLhcoreClassModelChatConfig::fetch('cleanup_cronjob')->current_value;
 
-        $timeoutCleanup = erLhcoreClassModelChatConfig::fetch('tracked_users_cleanup')->current_value;
+        if ($cleanupCronjob == 0 || (isset($params['cronjob']) && $params['cronjob'] == 1))
+        {
+            $lastCleanup = erLhcoreClassModelChatConfig::fetch('tracked_users_cleanup_last');
 
-        $lastCleanup = erLhcoreClassModelChatConfig::fetch('tracked_users_cleanup_last');
+            // Do not clean more often that once per hour
+            if ((int)$lastCleanup->current_value < time()-3600) {
 
-        // Do not clean more often that once per hour
-        if ((int)$lastCleanup->current_value < time()-3600) {
+                $timeoutCleanup = erLhcoreClassModelChatConfig::fetch('tracked_users_cleanup')->current_value;
 
-            $lastCleanup->identifier = 'tracked_users_cleanup_last';
-            $lastCleanup->type = 0;
-            $lastCleanup->explain = 'Track last cleanup';
-            $lastCleanup->hidden = 1;
-            $lastCleanup->value = time();
-            $lastCleanup->saveThis();
+                $timeoutCleanupFootprint = erLhcoreClassModelChatConfig::fetch('tracked_footprint_cleanup')->current_value;
 
-            $stmt = $db->prepare('DELETE T2 FROM lh_abstract_proactive_chat_event as T2 INNER JOIN lh_chat_online_user as T1 ON T1.id = T2.vid_id WHERE last_visit < :last_activity');
-            $stmt->bindValue(':last_activity', (int)(time() - $timeoutCleanup * 24 * 3600), PDO::PARAM_INT);
-            $stmt->execute();
+                $lastCleanup->identifier = 'tracked_users_cleanup_last';
+                $lastCleanup->type = 0;
+                $lastCleanup->explain = 'Track last cleanup';
+                $lastCleanup->hidden = 1;
+                $lastCleanup->value = time();
+                $lastCleanup->saveThis();
 
-            $stmt = $db->prepare('DELETE FROM lh_chat_online_user WHERE last_visit < :last_activity');
-            $stmt->bindValue(':last_activity', (int)(time() - $timeoutCleanup * 24 * 3600), PDO::PARAM_INT);
-            $stmt->execute();
+                $db = ezcDbInstance::get();
 
-            $stmt = $db->prepare('DELETE FROM lh_chat_online_user_footprint WHERE chat_id = 0 AND vtime < :last_activity');
-            $stmt->bindValue(':last_activity', (int)(time() - $timeoutCleanup * 24 * 3600), PDO::PARAM_INT);
-            $stmt->execute();
+                if ($timeoutCleanup > 0) {
+                    // Proactive events cleanup
+                    $stmt = $db->prepare('DELETE T2 FROM lh_abstract_proactive_chat_event as T2 INNER JOIN lh_chat_online_user as T1 ON T1.id = T2.vid_id WHERE last_visit < :last_visit');
+                    $stmt->bindValue(':last_visit', (int)(time() - ($timeoutCleanup * 24 * 3600)), PDO::PARAM_INT);
+                    $stmt->execute();
 
+                    // Online user cleanup
+                    $stmt = $db->prepare('DELETE FROM lh_chat_online_user WHERE last_visit < :last_activity');
+                    $stmt->bindValue(':last_activity', (int)(time() - ($timeoutCleanup * 24 * 3600)), PDO::PARAM_INT);
+                    $stmt->execute();
+                }
 
+                if ($timeoutCleanupFootprint > 0) {
+                    // Footprint
+                    $stmt = $db->prepare('DELETE FROM lh_chat_online_user_footprint WHERE vtime < :vtime LIMIT 50000');
+                    $stmt->bindValue(':vtime', (int)(time() - ($timeoutCleanupFootprint * 24 * 3600)), PDO::PARAM_INT);
+                    $stmt->execute();
+                }
+            }
         }
     }
